apps/vote_service/requirements.txt
Django
djangorestframework
redis
celery
psycopg2-binary
pika
django-cors-headers